Refactor: ThemeForm 삭제 후 병합
